pyproject.toml
[project]
name = "python-game-jam-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0",
    "pygame-ce>=2.5.7",
    "pytmx>=3.32",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
    "ty>=0.0.24",
]
